+++ views/legend.php
@@ -4,11 +4,11 @@
 
         <?php foreach ( $teccc->terms as $id => $attributes ): ?>
             <?php
-            $slug = esc_attr($attributes[Tribe_Events_Category_Colors::SLUG]);
-            $name = esc_attr($attributes[Tribe_Events_Category_Colors::NAME]);
-            $link = get_term_link($id, TribeEvents::TAXONOMY);
+            $slug = esc_attr( $attributes[Tribe_Events_Category_Colors::SLUG ] );
+            $name = esc_attr( $attributes[ Tribe_Events_Category_Colors::NAME ] );
+            $link = get_term_link( $id, TribeEvents::TAXONOMY );
             ?>
-            <li class="tribe-events-category-<?php echo $slug ?> tribe-events-category-<?php esc_attr_e($id) ?>">
+            <li class="tribe-events-category-<?php echo $slug ?> tribe-events-category-<?php esc_attr_e( $id ) ?>">
                 <a href="<?php echo $link ?>">
                     <?php echo $name ?>
                 </a>