+++ CHANGES.md
@@ -1,6 +1,7 @@
 
 #### 3.9.0
 * now using generated CSS stylesheet for events pages - yeah @barryhughes
+* no longer use inline styles, using generated stylesheet.
 
 #### 3.8.0
 * added coloring to Filter Bar in checkbox mode